+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Write the date in place of the "Unreleased" in the case a new version is release
 # Changelog
 
 
+### Added
+
+- A new "revoke:apikeys:self" scope that allows a user to revoke only the API key they are currently authenticated with.
+
+
 ## v0.2.12 (2026-06-16)
 
 ### Added